[refactor] 상품 상세 페이지 스타일드 컴포넌트 분리
