+++ client/src/pages/ItemDetail.js
@@ -1,125 +1,37 @@
-import styled from 'styled-components';
 import { useState, useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useNavigate, useLocation } from 'react-router-dom';
 import { useCookies } from 'react-cookie';
-import newAxios from '../components/newAxios';
+import jwt_decode from 'jwt-decode';
+import instance from '../components/newAxios';
 
 import Main from '../components/main/Main';
 import MainLayout from '../components/main/MainLayout';
 import Footer from '../components/main/Footer';
-
-import { FaShoppingCart } from 'react-icons/fa';
+import {
+  ImageBox,
+  Modal,
+  ProductDescription,
+  ProductInformation,
+  ReviewContainer,
+} from '../components/Item-detail/ItemDetail.styled';
 import ReviewForm from '../components/review/ReviewForm';
 import ReadReviews from '../components/review/ReadReviews';
-
-import {
-  findProductByProductId,
-  getProductReviews,
-} from '../components/api/itemDetailAPI';
-
-const ImageBox = styled.div`
-  width: 100%;
-  padding: 30px 0;
-  > img {
-    width: 100%;
-    height: 300px;
-    object-fit: cover;
-  }
-`;
-
-const ProductInformation = styled.div`
-  display: flex;
-  flex-direction: row;
-  > button {
-    width: 20%;
-  }
-  > div {
-    width: 80%;
-    word-break: keep-all;
-    #product-name {
-      font-size: xx-large;
-    }
-    #product-price {
-      font-size: xx-large;
-    }
-  }
-  > icon {
-    width: 20%;
-  }
-`;
-
-const ProductDescription = styled.p`
-  margin: 10px 0 50px 0;
-  word-break: break-all;
-  color: ${(props) => props.color};
-  img {
-    width: 500px;
-  }
-`;
-
-const Modal = styled.div`
-  position: fixed;
-  bottom: 300px;
-  right: 0;
-  transform: translate(-50%, 0);
-  z-index: 999;
-  max-width: 300px;
-  min-height: 150px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--white);
-  border: 1px solid var(--border);
-  border-radius: var(--bd-rd);
-  > div {
-    > div {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      > button {
-        margin-top: 30px;
-        width: 150px;
-        height: 40px;
-        border-radius: var(--bd-rd);
-        background-color: var(--border);
-        color: var(--black);
-      }
-    }
-  }
-`;
-
-const ReviewContainer = styled.div`
-  > h3 {
-    margin: 30px 16px;
-    padding: 16px 0;
-    border-bottom: 2px solid var(--border);
-  }
-`;
-
-const FormButton = styled.button`
-  width: 100px;
-  height: 40px;
-  border-radius: var(--bd-rd);
-  background-color: ${(props) => props.backgroundColor};
-  color: var(--white);
-  margin-left: 10px;
-`;
+import AdminButtonGroup from '../components/Item-detail/AdminButtonGroup';
+import { FaShoppingCart } from 'react-icons/fa';
 
 const ItemDetail = () => {
-  const [cookies] = useCookies();
-
-  const [productDetail, setProductDetail] = useState(null);
-  const [productReviews, setProductReviews] = useState(null);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isEditClicked, setIsEditClicked] = useState(false);
   const [editingReview, setEditingReview] = useState(null);
+  const [userRole, setUserRole] = useState('');
 
-  const refreshToken = cookies.refreshToken;
+  const [cookies] = useCookies();
 
   const navigate = useNavigate();
 
-  const { id } = useParams();
+  const location = useLocation();
+  const productDetail = location.state.responseProductDetail;
+  const productReviews = location.state.responseReviews;
 
   const handleModal = () => {
     setIsModalOpen((isOpen) => !isOpen);
@@ -127,39 +39,33 @@ const ItemDetail = () => {
   };
 
   const handleShoppingBag = () => {
+    if (userRole === '') {
+      window.alert('상품을 장바구니에 담기 전에 로그인 해주세요.');
+      return navigate('/login');
+    }
+
     const item = {
       productId: productDetail.productId,
       quantity: 1,
     };
 
-    newAxios
+    instance
       .post(`/carts`, item)
       .then(() => setIsModalOpen(true))
       .catch((error) => {
         if (error.response.data.status === 409) {
-          window.alert('이미 장바구니에 담은 상품입니다.');
+          alert('이미 장바구니에 담은 상품입니다.');
           navigate('/shoppingcart');
         }
       });
   };
 
-  const handleEditProductDetail = () => {
-    navigate(`/admin-item/${id}`, { state: productDetail });
-  };
-
-  const handleDeleteProductDetail = () => {
-    newAxios.delete(`../products/${productDetail.productId}`);
-    navigate('/product');
-  };
-
   useEffect(() => {
-    findProductByProductId(id).then((productInformation) => {
-      setProductDetail(productInformation);
-    });
-    getProductReviews(id).then((productReviewList) =>
-      setProductReviews(productReviewList)
-    );
-  }, [id]);
+    if (Object.keys(cookies).length) {
+      const { accessToken } = cookies;
+      setUserRole(jwt_decode(accessToken).roles[0]);
+    }
+  }, [cookies]);
 
   return (
     <Main>
@@ -179,28 +85,11 @@ const ItemDetail = () => {
                   {productDetail.price.toLocaleString('ko-KR')}원
                 </div>
               </div>
-              {refreshToken && (
-                <button type="button" onClick={handleShoppingBag}>
-                  <FaShoppingCart size="40px" />
-                </button>
-              )}
-              {refreshToken && (
-                <>
-                  <FormButton
-                    type="button"
-                    backgroundColor="#61a0ff"
-                    onClick={handleEditProductDetail}
-                  >
-                    상품 수정하기
-                  </FormButton>
-                  <FormButton
-                    type="button"
-                    backgroundColor="#ff0000"
-                    onClick={handleDeleteProductDetail}
-                  >
-                    상품 삭제하기
-                  </FormButton>
-                </>
+              <button type="button" onClick={handleShoppingBag}>
+                <FaShoppingCart size="40px" />
+              </button>
+              {userRole === 'ADMIN' && (
+                <AdminButtonGroup productDetail={productDetail} />
               )}
               {isModalOpen && (
                 <Modal>
@@ -217,7 +106,7 @@ const ItemDetail = () => {
               {productDetail.subtitle}
             </ProductDescription>
             <ProductDescription
-              dangerouslySetInnerHTML={{ __html: productDetail.productDetail }}
+              dangerouslySetInnerHTML={{ __html: productDetail.productDetail }} // html 그대로 사용하지 않는 방법 연구하기
             ></ProductDescription>
           </div>
         )}
